feat(auth): add lib/auth/session.ts with token ref and types
